API/requirements.txt
django
djangorestframework
environs
axios